+++ src/Extensions/InfrastructureExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.IO;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using ViajaNet.JobApplication.Infrastructure.CouchDb;
 using ViajaNet.JobApplication.Infrastructure.Queue;
 
 namespace ViajaNet.JobApplication.Extensions
@@ -8,6 +9,8 @@ namespace ViajaNet.JobApplication.Extensions
     public static class InfrastructureExtensions
     {
         private const string QueueConfigurationSection = "PubSub:RabbitMq";
+        private const string CouchDbConfigurationSection = "Repository:CouchDb";
+        private const string SqlServerConfigurationSection = "Repository:SqlServer";
 
         public static IServiceCollection AddViajaNetProviders(this IServiceCollection services)
         {
@@ -28,15 +31,21 @@ namespace ViajaNet.JobApplication.Extensions
             return services.AddOptions()
                 .Configure<QueueConfiguration>(options => configuration.GetSection(QueueConfigurationSection).Bind(options))
                 .AddSingleton<IQueueFactory, QueueProviderFactory>()
-                .AddSingleton<IQueueProvider, QueueService>();
+                .AddTransient<IQueueProvider, QueueService>();
         }
 
-        public static IServiceCollection AddViajaNetRepositories(this IServiceCollection services, IConfiguration configuration)
+        public static IServiceCollection AddViajaNetRepositories(this IServiceCollection services)
         {
+            var configBuilder = new ConfigurationBuilder()
+                                    .SetBasePath(Directory.GetCurrentDirectory())
+                                    .AddJsonFile($"appsettings.json", optional: true);
+
+            var configuration = configBuilder.Build();
+
             return services.AddOptions()
-                .Configure<QueueConfiguration>(options => configuration.GetSection(QueueConfigurationSection).Bind(options))
-                .AddSingleton<IQueueFactory, QueueProviderFactory>()
-                .AddSingleton<IQueueProvider, QueueService>();
+                .Configure<CouchDbConfiguration>(options => configuration.GetSection(CouchDbConfigurationSection).Bind(options))
+                .AddSingleton<ICouchDbFactory, CouchDbFactory>()
+                .AddTransient<ICouchDbService, CouchDbService>();
         }
     }
 }