+++ src/Infrastructure/CouchDb/ICouchDbFactory.cs
@@ -1,0 +1,10 @@
+using System;
+using Couchbase.Core;
+
+namespace ViajaNet.JobApplication.Infrastructure.CouchDb
+{
+    public interface ICouchDbFactory : IDisposable
+    {
+        IBucket CreateBucket();
+    }
+}