+++ src/Infrastructure/Queue/IQueueFactory.cs
@@ -1,8 +1,9 @@
+using System;
 using RabbitMQ.Client;
 
 namespace ViajaNet.JobApplication.Infrastructure.Queue
 {
-    public interface IQueueFactory
+    public interface IQueueFactory : IDisposable
     {
         IModel CreateChannel();
     }