+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2020-01-03
+### Added
+- CouchDb factory and service implementation;
+- Added configuration data on appSettings.json.
+
+### Changed
+- Refactored providers and services to supress garbage collector on disposal.
+
 ## [0.6.0] - 2020-01-03
 ### Added
 - Controller's application service dependency injection;